Fix tests and rename Integrate to Contract
