+++ funsor/__init__.py
@@ -6,7 +6,7 @@ from funsor.terms import Funsor, Number, Variable, of_shape, to_funsor
 from funsor.torch import Function, Tensor, arange, function, torch_einsum
 
 from . import (adjoint, contract, delta, distributions, domains, einsum, gaussian, handlers, interpreter, joint,
-               minipyro, ops, terms, torch)
+               minipyro, ops, sum_product, terms, torch)
 
 __all__ = [
     'Domain',
@@ -35,6 +35,7 @@ __all__ = [
     'ops',
     'reals',
     'reinterpret',
+    'sum_product',
     'terms',
     'to_funsor',
     'torch',
