+++ test/test_contract.py
@@ -1,81 +1,80 @@
 from __future__ import absolute_import, division, print_function
 
-from collections import OrderedDict
-
+from collections import OrderedDict  # noqa: F401
 import pytest
-from six.moves import reduce
+import torch  # noqa: F401
+
+import funsor
 
 import funsor.ops as ops
-from funsor.contract import _partition, partial_sum_product, sum_product
-from funsor.domains import bint
-from funsor.testing import assert_close, random_tensor
+from funsor.domains import bint  # noqa: F401
+from funsor.interpreter import interpretation, reinterpret
+from funsor.optimizer import Finitary, optimize
+from funsor.terms import reflect
+from funsor.testing import assert_close, make_einsum_example
+from funsor.torch import Tensor  # noqa: F401
+
+from funsor.einsum import einsum
+from funsor.contract import Contract, naive_contract_einsum
 
 
-@pytest.mark.parametrize('inputs,dims,expected_num_components', [
-    ([''], set(), 1),
-    (['a'], set(), 1),
-    (['a'], set('a'), 1),
-    (['a', 'a'], set(), 2),
-    (['a', 'a'], set('a'), 1),
-    (['a', 'a', 'b', 'b'], set(), 4),
-    (['a', 'a', 'b', 'b'], set('a'), 3),
-    (['a', 'a', 'b', 'b'], set('b'), 3),
-    (['a', 'a', 'b', 'b'], set('ab'), 2),
-    (['a', 'ab', 'b'], set(), 3),
-    (['a', 'ab', 'b'], set('a'), 2),
-    (['a', 'ab', 'b'], set('b'), 2),
-    (['a', 'ab', 'b'], set('ab'), 1),
-    (['a', 'ab', 'bc', 'c'], set(), 4),
-    (['a', 'ab', 'bc', 'c'], set('c'), 3),
-    (['a', 'ab', 'bc', 'c'], set('b'), 3),
-    (['a', 'ab', 'bc', 'c'], set('a'), 3),
-    (['a', 'ab', 'bc', 'c'], set('ac'), 2),
-    (['a', 'ab', 'bc', 'c'], set('abc'), 1),
+EINSUM_EXAMPLES = [
+    "a,b->",
+    "ab,a->",
+    "a,a->",
+    "a,a,a,ab->",
+    "ab->",
+    "ab,bc,cd->",
+    "abc,bcd,def->",
+    "abc,abc,bcd,bcd,def,def->",
+    "ab,bc,cd,de->",
+    "ab,ab,bc,bc,cd,cd->",
+]
+
+
+@pytest.mark.parametrize('equation', EINSUM_EXAMPLES)
+@pytest.mark.parametrize('backend,fill', [
+    ('torch', None),
+    ('torch', 1.),
+    ('pyro.ops.einsum.torch_log', None)
 ])
-def test_partition(inputs, dims, expected_num_components):
-    sizes = dict(zip('abc', [2, 3, 4]))
-    terms = [random_tensor(OrderedDict((s, bint(sizes[s])) for s in input_))
-             for input_ in inputs]
-    components = list(_partition(terms, dims))
+def test_contract_einsum_product_measure(equation, backend, fill):
+    inputs, outputs, sizes, operands, funsor_operands = make_einsum_example(equation, fill=fill)
 
-    # Check that result is a partition.
-    expected_terms = sorted(terms, key=id)
-    actual_terms = sorted((x for c in components for x in c[0]), key=id)
-    assert actual_terms == expected_terms
-    assert dims == set.union(set(), *(c[1] for c in components))
+    with interpretation(reflect):
+        expected = einsum(equation, *funsor_operands, backend=backend)
+        print("TRUE GRAPH: {}".format(expected))
+    expected = reinterpret(expected)
+    actual = naive_contract_einsum(equation, *funsor_operands, backend=backend)
 
-    # Check that the partition is not too coarse.
-    assert len(components) == expected_num_components
-
-    # Check that partition is not too fine.
-    component_dict = {x: i for i, (terms, _) in enumerate(components) for x in terms}
-    for x in terms:
-        for y in terms:
-            if x is not y:
-                if dims.intersection(x.inputs, y.inputs):
-                    assert component_dict[x] == component_dict[y]
+    assert isinstance(actual, funsor.Tensor) and len(outputs) == 1
+    print(expected / actual, actual / expected)
+    assert_close(expected, actual, atol=1e-4)
+    for output in outputs:
+        for i, output_dim in enumerate(output):
+            assert output_dim in actual.inputs
+            assert actual.inputs[output_dim].dtype == sizes[output_dim]
 
 
-@pytest.mark.parametrize('sum_op,prod_op', [(ops.add, ops.mul), (ops.logaddexp, ops.add)])
-@pytest.mark.parametrize('inputs,plates', [('a,abi,bcij', 'ij')])
-@pytest.mark.parametrize('vars1,vars2', [
-    ('', 'abcij'),
-    ('c', 'abij'),
-    ('cj', 'abi'),
-    ('bcj', 'ai'),
-    ('bcij', 'a'),
-    ('abcij', ''),
-])
-def test_partial_sum_product(sum_op, prod_op, inputs, plates, vars1, vars2):
-    inputs = inputs.split(',')
-    factors = [random_tensor(OrderedDict((d, bint(2)) for d in ds)) for ds in inputs]
-    plates = frozenset(plates)
-    vars1 = frozenset(vars1)
-    vars2 = frozenset(vars2)
+@pytest.mark.parametrize('equation1', EINSUM_EXAMPLES)
+@pytest.mark.parametrize('equation2', EINSUM_EXAMPLES)
+def test_contract_naive_pair(equation1, equation2):
 
-    factors1 = partial_sum_product(sum_op, prod_op, factors, vars1, plates)
-    factors2 = partial_sum_product(sum_op, prod_op, factors1, vars2, plates)
-    actual = reduce(prod_op, factors2)
+    # identical structure
+    funsor_operands1 = [a.abs() for a in make_einsum_example(equation1)[-1]]
+    funsor_operands2 = [a.abs() for a in make_einsum_example(equation2)[-1]]
 
-    expected = sum_product(sum_op, prod_op, factors, vars1 | vars2, plates)
-    assert_close(actual, expected)
+    with interpretation(optimize):
+        measure = Finitary(ops.mul, tuple(funsor_operands1))
+        integrand = Finitary(ops.mul, tuple(funsor_operands2))
+
+        intermediate = measure * integrand
+        expected = intermediate.reduce(ops.add)
+
+        actual = Contract(measure, integrand,
+                           frozenset(measure.inputs) | frozenset(integrand.inputs))
+
+    actual = reinterpret(actual)
+    expected = reinterpret(expected)
+
+    assert_close(expected, actual, atol=1e-4)
